pyproject.toml
[project]
name = "AxisMD"
version = "0.1.0"
description = ""
authors = [
    {name = "UmairAsDev",email = "umairashrafbsse@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "postgres (>=4.0,<5.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "openai (>=1.99.9,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyprojroot (>=0.3.0,<0.4.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "noisereduce (>=3.0.3,<4.0.0)",
    "speechrecognition (>=3.14.3,<4.0.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "webrtcvad (>=2.0.10,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.31,<0.4.0)",
    "langchain-core (>=0.3.74,<0.4.0)",
    "whisper-openai (>=1.0.0,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "sounddevice (>=0.5.2,<0.6.0)",
    "langsmith (>=0.4.16,<0.5.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
